Andy Ross
The biggest change is that I
reduced the cruise speed (360->310) to represent what a typical
aircraft "should" be able to achieve with 50% fuel. The resulting
aircraft is a *lot* less slick (it produces almost three times as much
parasite drag), and should perform more realistically at low
altitudes. You should still be able to hit the textbook numbers with
a stripped aircraft and empty tanks.

Also, there is a syntax fix: the <piston-engine> tag should be a
subtag of the <propeller> tag. Putting it immediately after the
propeller definition works only due to a lucky interaction with the
way the parser stores its "current object" pointer. It doesn't clear
the value when it sees </propeller>, so you can get away with putting
the engine at any point in the file before the next tag that sets a
current object.