hvengel on the forum: The most glaring issue is twist="1.0" which would create a wing with wash in instead of wash out. No real world aircraft has wash in since this will result in a plane that is very unstable. This should be twist="-1.0" so that the wing is modeled as having wash out.
In addition I was able to find a web page that has the data for the root section of the p51's airfoil: http://www.worldofkrauss.com/foils/1085

Based on the data there I changed the camber setting to 0.123 and the <stall> settings to

<stall aoa="10.5" width="6" peak="1.5"/>

Camber was calculated based on what the YASim docs say using the airfoil data from the web page. Hint - YASim camber has nothing to do with the airfoil camber. The aoa setting is right out of the airfoil data but width and peak are nothing more than guesses although peak is based on the YASim docs and the airfoil data. These do result in the plane flying OK but the stall may be a bit too abrupt and it may need to have width increased a little. The following is what the <wing> section of my XML file looks like now.

Complete story:
http://flightgear.org/forums/viewtopic.php?f=4&t=6734