hvengel: It turns out that Models/p512d-jw.xml had some settings for the right left and tail gear mixed (IE. connected to the wrong properties).
A very small change to the air speed indicator so that it give a little more feed back below 100MPH.

The last patch results in a flyable P51D that has performance that approximates the real thing. Ground handling is still a little tricky and it may be advisable to move the wing back a little farther. Stalls are abrupt and it will snap to the left if you do not react to the stall as soon as there is the slightest indication that one is developing. There is very little warning that a stall is about to happen. I am not sure if the stalls are currently too abrupt relative to the real thing but from my reading about how the real thing flies I think it is about right. You do not want to stall on approach as you will auger in. In addition accelerated stall are viscous and result in a spin.